Card comparisons and stuff

+++ game-engine/src/main/java/de/poker/trainer/engine/Player.java
@@ -1,11 +1,11 @@
 package de.poker.trainer.engine;
 
-public record Player(Position position, int stack) {
-    public static Player pay(Player player, int amount) {
+public record Player(Position position, double stack) {
+    public static Player pay(Player player, double amount) {
         return new Player(player.position, player.stack - amount);
     }
 
-    public static Player winPot(Player player, int potSize) {
+    public static Player winPot(Player player, double potSize) {
         return new Player(player.position, player.stack() + potSize);
     }
 }
